+++ plugins/postcss-nesting/mod.js
@@ -1,3 +1,3 @@
 /* Deno entry point */
 
-export { default } from './dist/index.deno.mjs'
+export { default } from './dist/index.deno.mjs';