+++ plugins/postcss-nesting/src/lib/merge-selectors/compound-selector-order.js
@@ -57,28 +57,29 @@ export function sortCompoundSelector(node) {
 	// `h1.foo`
 
 	node.nodes.sort((a, b) => {
-		if (a.type === b.type) {
-			return 0;
-		}
-
-		if (selectorTypeOrder[a.type] < selectorTypeOrder[b.type]) {
-			return -1;
-		}
-
-		return 1;
+		return selectorTypeOrder(a, a.type) - selectorTypeOrder(b, b.type);
 	});
 }
 
-const selectorTypeOrder = {
+function selectorTypeOrder(selector, type) {
+	if (parser.isPseudoElement(selector)) {
+		return selectorTypeOrderIndex.pseudoElement;
+	}
+
+	return selectorTypeOrderIndex[type];
+}
+
+const selectorTypeOrderIndex = {
 	universal: 0,
 	tag: 1,
 	id: 2,
 	class: 3,
 	attribute: 4,
-	pseudo: 5,
-	selector: 7,
+	selector: 5,
+	pseudo: 6,
+	pseudoElement: 7,
 	string: 8,
-	root : 9,
+	root: 9,
 	comment: 10,
 
 	nesting: 9999,