+++ plugins/postcss-oklab-function/src/convert-oklab-to-srgb.ts
@@ -11,7 +11,7 @@ export function oklabToSRgb(oklabRaw: color): color {
 		0,
 	);
 
-	const oklab = [oklabL / 100, oklabARaw, oklabBRaw] as color;
+	const oklab = [oklabL, oklabARaw, oklabBRaw] as color;
 
 	let conversion = oklab as color;
 	let oklch = OKLab_to_OKLCH(conversion);