+++ rollup/configs/externals.js
@@ -3,14 +3,18 @@ export const externalsForCLI = [
 	'path',
 	'url',
 
+	'@csstools/postcss-cascade-layers',
 	'@csstools/postcss-color-function',
 	'@csstools/postcss-font-format-keywords',
-	'@csstools/postcss-ic-unit',
 	'@csstools/postcss-hwb-function',
+	'@csstools/postcss-ic-unit',
 	'@csstools/postcss-is-pseudo-class',
 	'@csstools/postcss-normalize-display-values',
 	'@csstools/postcss-oklab-function',
 	'@csstools/postcss-progressive-custom-properties',
+	'@csstools/postcss-stepped-value-functions',
+	'@csstools/postcss-unset-value',
+	'@csstools/selector-specificity',
 	'autoprefixer',
 	'browserslist',
 	'caniuse-lite',
@@ -46,7 +50,6 @@ export const externalsForCLI = [
 	'postcss-selector-not',
 	'postcss-selector-parser',
 	'postcss-value-parser',
-	'postcss-values-parser',
 ];
 
 export const externalsForPlugin = [
@@ -59,6 +62,7 @@ export const externalsForPlugin = [
 	/^postcss-\d\.\d$/,
 	/^postcss\/lib\/*/,
 
+	'@csstools/postcss-cascade-layers',
 	'@csstools/postcss-color-function',
 	'@csstools/postcss-font-format-keywords',
 	'@csstools/postcss-hwb-function',
@@ -67,6 +71,9 @@ export const externalsForPlugin = [
 	'@csstools/postcss-normalize-display-values',
 	'@csstools/postcss-oklab-function',
 	'@csstools/postcss-progressive-custom-properties',
+	'@csstools/postcss-stepped-value-functions',
+	'@csstools/postcss-unset-value',
+	'@csstools/selector-specificity',
 	'autoprefixer',
 	'browserslist',
 	'caniuse-lite',
@@ -104,7 +111,6 @@ export const externalsForPlugin = [
 	'postcss-selector-not',
 	'postcss-selector-parser',
 	'postcss-value-parser',
-	'postcss-values-parser',
 ];
 
 export const externalsForBrowser = [];