+++ plugins/postcss-is-pseudo-class/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changes to PostCSS Is Pseudo Class
 
+### 2.0.4 (May 17, 2022)
+
+ - Fix selector order with `:before` and other pseudo elements.
+
+### 2.0.3 (May 11, 2022)
+
+- Use `@csstools/selector-specificity` for specificity calculations.
+
+### 2.0.2 (April 4, 2022)
+
+- Improved : compound selector order with pseudo elements
+- Improved : selector specificity calculation.
+
 ### 2.0.1 (March 4, 2022)
 
 - Preserve selector order as much as possible. Fixes issues where pseudo elements `::before` were moved.